pyproject.toml
[project]
name = "prl-bib2html"
version = "0.1.0"
description = "Python library for converting BibTeX files to HTML for academic publications websites"
requires-python = ">=3.8"
dependencies = [
    "bibtexparser",
    "requests",
    "pyyaml"
]

[project.optional-dependencies]
demo = [
    "Flask"
]

[project.scripts]
prl-bib2html = "prl_bib2html.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["prl_bib2html*"]